Updated the migration

+++ PPMTool/Migrations/20250430065548_CullSkills.cs
@@ -10,35 +10,33 @@ namespace PPMTool.Migrations
         /// <inheritdoc />
         protected override void Up(MigrationBuilder migrationBuilder)
         {
-            // Read from the data file and remove the skills with controlled names that are in the file
-            migrationBuilder.Sql(@"
-                -- Create a temporary table to store the skills to cull
-                CREATE TEMP TABLE SkillsToCull (SkillName TEXT);
+            // Read skills from file
+            var localFilePath = "./Migrations/Data/SkillsToCull.txt";
+            if (!File.Exists(localFilePath)) throw new FileNotFoundException("File not found", localFilePath);
+            var skillsToCull = File.ReadAllLines(localFilePath);
 
-                -- Read the SkillsToCull.txt file and insert each line into the temporary table
-                INSERT INTO SkillsToCull (SkillName)
-                SELECT TRIM(LOWER(value))
-                FROM readfile('SkillsToCull.txt');
+            // Remove the skill tag from tasks and people and then delete from table
+            foreach (var skill in skillsToCull)
+            {
+                migrationBuilder.Sql(@$"
+                    DELETE FROM OwnedSkills
+                    WHERE SkillId IN (
+                        SELECT SkillId
+                        FROM SkillTags
+                        WHERE LOWER(ControlledName) == LOWER({skill})
+                    );
 
-                -- Iterate over each skill to cull
-                DELETE FROM OwnedSkills
-                WHERE SkillId IN (
-                    SELECT SkillId
-                    FROM SkillTags
-                    WHERE LOWER(ControlledName) IN (SELECT SkillName FROM SkillsToCull)
-                );
+                    DELETE FROM SkillTagSubTask
+                    WHERE SkillsRequiredSkillTagId IN (
+                        SELECT SkillId
+                        FROM SkillTags
+                        WHERE LOWER(ControlledName) == LOWER({skill})
+                    );
 
-                DELETE FROM SkillTagSubTask
-                WHERE SkillsRequiredSkillTagId IN (
-                    SELECT SkillId
-                    FROM SkillTags
-                    WHERE LOWER(ControlledName) IN (SELECT SkillName FROM SkillsToCull)
-                );
-
-                -- Drop the temporary table
-                DROP TABLE SkillsToCull;
-                )   
-            ");
+                    DELETE FROM SkillTags
+                    WHERE LOWER(ControlledName) == LOWER({skill});
+                ");
+            }
         }
 
         /// <inheritdoc />
